Fix enchant_broker_list_dicts
List the top priority dictionary for each language tag.

Note that this lists real dictionaries, but may not list the provider and
dictionary that would actually be used for a given tag: for example, if
myspell is the preferred provider for tag “en”, and it has an “en_UK”
dictionary, it would be used to check “en” even if aspell, say, provides
exactly an “en” dictionary. However, the dictionary list would show the
aspell dictionary.

See issue #8 for more details.