Don’t deduplicate module & configuration paths
This change was introduced in commit fb312be to stop enchant-lmod listing
duplicate modules. However, it makes an undocumented change to the way that
configuration paths work, and can give surprising results: if a
higher-priority configuration source happens to use the same path as a
lower-priority one, then it will be ignored.

Since a side effect of the changes to enchant_broker_list_dicts in commit
854316a is to make it list only one dictionary per tag in any case, this
uniquifying is no longer needed, so there is no corresponding change in
enchant-lsmod.c.