Make library relocatable using gnulib’s relocatable-lib-lgpl module
This adds a new API enchant_set_prefix_dir, which necessitated changing the
C++ API so that rather than constructing a broker instance statically, which
would pre-empt any setting of the prefix dir, the Broker object must now be
explicitly created. This also gives an easier way to have multiple Brokers,
if desired.