Require providers to define dispose methods, and one of dictionary_exists or list_dicts
Both dispose and dispose_dicts are required. This makes it harder to have
inadvertent leaks, and simplifies lib.c a little.

Require one of dictionary_exists or list_dicts. This means that request_dict
is no longer called by enchant_provider_dictonary_exists, so remove the
tests for this case.