Get AppleSpell back-end working again, and add list_dicts method
It transpired that owing to a problem in configure.ac, the backend was
not being built, even on macOS. Fix this.

As a result, some compilation errors recently introduced emerged. Fix
them.

Note that AppleSpell.config wasn't being installed. Fix this. Also fix
its loading: it was being looked for in pkglibdir, but should be
installed in pkgdatadir (and now is).

In order to find the config file, promote enchant_get_conf_dirs to
enchant-provider.h, which now therefore needs to include glib.h again.