Overhaul hspell back-end
Use C99-style declarations.

Remove check of whether text to be checked is in Hebrew, as hspell already
does this (and in fact it’s not what we want: words in non-Hebrew are
treated as “empty” and therefore correct; this will have to be dealt with by
having the Enchant back-end reject words not in Hebrew, but probably it’s
better to have generic code to do this which detects words that contain
non-word characters for the given dictionary; however, that will require the
implementation of issue #17).