configure.ac: change definition of shared library’s AGE
Micro version, according to the rules in configure.ac, seems to match the
definition given in libtool’s manual better, which is a little odd, because
libvirt, from which I copied the version-numbering code, is focused on
matching libtool’s definitions.

But here are the rules for AGE:

 5. If any interfaces have been added since the last public release,
     then increment AGE.

  6. If any interfaces have been removed or changed since the last
     public release, then set AGE to 0.

This exactly matches the rules for the minor version number in configure.ac:

     4a) 5) Increment minor when adding interfaces.
     6) Set minor to zero when removing or changing interfaces.

Further, adding 1000 * MAJOR_VERSION to AGE, as previously, seems to
conflict with rule 6, above, for AGE.

This should also resolve issue #153 to everyone’s satisfaction.