Compute libdir_subdir for the tests properly (fixes issue #185)
This is potentially nasty: the user who reported the issue noticed that the
tests tried to delete their system libraries!

Unfortunately, this requires setting a make variable from a shell
command (unless anyone has a better idea?).

This is not possible in POSIX make.

There is a suggestion to make it possible using the != assignment syntax,
which is already widely supported, but unfortunately automake rejects that
syntax.

Therefore, use $(shell), and use the gnu-make gnulib module to detect GNU
Make; if it is not being used, try setting libdir_subdir=lib, which will
work with default settings.

Document this requirement in INSTALL.