Avoid trailing slash in `rm -rf $(var)/'
When removing via `rm -rf' adding a trailing slash can be dangerous
when `$(var)' is undefined and ends up in completely removing `/'
as happened in #185 and #201.
They are not needed and it is better to avoid them.

This avoid `rm -rf /' when running `make check'.

Please note that if the current user has permission to write a possible
`/enchant-@ENCHANT_MAJOR_VERSION@' directory that directory will be created
(i.e. the problem described by issue #201 is not completely addressed but this
commit but should avoid possible nasty `rm -rf')