src/enchant.c: overhaul command-line parsing and documentation
Use getopt for simpler parsing of command-line arguments and better error
handling and reporting.

Remove some stream variables, as in many cases we always use stdout.

Overhaul the documentation of command-line options and make it consistent
between the output of ‘enchant -h’ and enchant(1).

Allow -L to be combined with another mode; it is useless otherwise!