tests: fix some space leaks
Accidentally trying to build with an unreleased git version of
automake (post 1.16.3) led to a problem with libtool configuration that, I
think, led to the test providers being built statically, and thereby
revealed more leaks, though a lot of the reported leaks seemed to be
bogus (leaks of C++ auto variables that should have been destroyed
automatically).

Also configure ASAN to report leak backtraces in full.