Remove support for provider-specific personal wordlists
Enchant’s support for provider-specific wordlists has been removed. Of the
known providers, only Aspell used this support, and only partially. This was
likely to be confusing: for example, words could be added to the Aspell, but
not removed, so Aspell’s personal wordlist could get out of sync with
Enchant’s. The relevant dictionary methods have been removed:
add_to_personal, add_to_exclude, and store_replacement.

A new dictionary method remove_from_session() allows providers to stay in
sync with Enchant’s personal wordlist, by removing as well as adding words
in the spelling session.

The public API enchant_dict_store_replacement() is deprecated, and no longer
does anything. It only previously had an effect with the Aspell provider.

Note that these changes do not affect the public ABI, which only uses
pointers to dictionary objects (it does change the provider ABI, but as far
as the maintainers know, there are no 3rd-party providers).