Translate PWL module into Vala
Also lift the limit on words in the PWL being at most BUFSIZ bytes long.
This seems a fairly arbitrary limit, some technical words can be 200k
letters long, and it’s not obvious what a better limit is, or that having an
indefinite limit will actually be a problem.