Remove Windows resource file
I don’t think this is really needed: as far as I can tell, Enchant is used
in environments that don’t require the extra metadata.

Also, allow the library to be versioned on Windows.

Removing these bits means that we no longer need to detect Windows at
configure time, and Enchant builds exactly the same regardless of system.