Make enchant_dict_suggest return an empty list for no suggestions
This makes it possible to distinguish an error from simply having no
suggestions.

Formally, this is a change to the API (I had to change some tests). But in
fact, it’s not a change, as this API could already return an empty list in
the case where the suggestions returned by the provider consisted entirely
of excluded words. Hence, I am not considering this to be an API/ABI break,
and am not bumping the major version, as callers already had to deal with
this eventuality. Further, calling code is unlikely to rely on the list
being non-empty.