Delay creation of EnchantDict to avoid partial initialization
To properly destruct an EnchantDict, its session field must be non-null since
the destructor assumes a session has been associated with the object. This
association happens in EnchantBroker when a provider upon request finds a valid
dictionary and in doing so creates an EnchantDict object which is returned to
the broker.

In the AppleSpell provider, the EnchantDict was created regardless of whether
the provider would be able to find a valid dictionary or not. If it did not
find a valid dictionary, it would erroneously free the EnchantDict object and
return null, despite the leaving the EnchantDict in the broker's set of
dictionaries.

This commit fixes the issue by moving the creation of the EnchantDict to occur
after successfully finding a valid dictionary.

Fixes #391