Fix word tokenization in enchant(1) (fix #412)
Thanks to @daeho-ro for the report, and apologies to users for this bad bug.

There were actually two bugs: the length of each word was off by one, and
the words after the first were not truncated when they should be. Hence,
only the first word was wrong in the case that failed.

Need to implement tests for the enchant program (#214)! I will add some
soon.