gnulib: fix build of relocatable.c on some Windows platforms
On some Windows platforms, command-line arguments that look like paths are automatically
transformed to Windows format.

This was happening with INSTALLDIR in gnulib's Makefile.am, but not INSTALLPREFIX, which
comes from config.h.

As a result, relocation was failing, as INSTALLPREFIX was not a prefix of INSTALLDIR.

To fix this, use configmake.h to get the path (BINDIR), so that it is expanded by the C
preprocessor, and not visible on the command line.