Allow provider ‘list_dicts’ and ‘suggest’ methods to return NULL strings
As a step towards coping better with memory allocation failures. In some
back-ends, the elements of the result arrays are created by [g_]strdup, and
checking whether each one is non-NULL and deallocating the array if so is
annoying and long-winded in C. So, since each list returned is filtered in
any case, simply add null pointer checks in the callers, and set an error
condition if one is found. The suggestions list returned will be NULL from
enchant_dict_suggest, and the provider will be skipped in
enchant_broker_list_dicts.

Document this in enchant-provider.h.