Make all dictionary methods formally optional
Actually, we want some of them to be always implemented, but for now check
them all explicitly.

Deal with PWL dictionaries this way, rather than with the special ‘is_pwl’
flag in the session, which is now removed.

Also fix a space leak in enchant(1) by changing the Enchant VAPI: it’s safe
for the return value of Dict.suggest() to be a regular array, which the Vala
caller can free. The C API has a special enchant_dict_free_string_list
method, but we can safely promise in the Vala API that it is a Vala array,
as the Vala implementation implements it thus.