Implement provider dictionaries as Dict classes, not delegates (WIP)
The parent EnchantDict class has been adjusted.

Next, need to add a C header file for the EnchantDict type, and a VAPI for
Enchant to use.

See https://docs.gtk.org/gobject/tutorial.html for how to write the header
file.

Then, use it in the providers to implement the class.