Implement provider dictionaries as subclasses of ProviderDict  (WIP)
The parent EnchantProviderDict class has been adjusted.

Next, need to add a C header file for the EnchantProviderDict type, and a
VAPI for Enchant to use.

See https://docs.gtk.org/gobject/tutorial.html for how to write the header
file.

Then, use it in the providers to implement the class.