enchant CLI: fix tokenization
The fix for issue #412 was incorrect in two ways (ironically, the number of
ways in which it claimed to fix the previous code), and it was a miracle
that it wasn’t causing segfaults.

The code to skip backwards over non-word characters at the end of a word did
nothing, because of a variable that was supposed to be a pointer to the word
but was in fact a copy of it. This means that a later attempt to truncate
the string used an entirely bogus length.

This bug did not appear when testing for #412 because the problem then was
that the tokenization code was truncating too much!

The second problem in the existing code was less serious: the code to remove
non-word characters at the end of the word was starting by considering the
character after the end of the word. This was not a real problem, because a)
there always is such a character (the input is NUL-terminated), and b) that
character is always a character that is not valid in a word in any position.
So the same result would have been obtained anyway; but still, fix this.

Since the #412 fix, we have now got a testing framework for enchant(1), so
add a test!